# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: BTL 2026 Transport API
  description: >
    Airport pickup and hotel shuttle coordination for Built to Last 2026 conference
    (June 11–14, Indianapolis).

    Admin endpoints require a JWT Bearer token obtained from `POST /api/v1/admin/login`.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local dev
  - url: https://api.btl2026.com/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /admin/login. Valid 12 hours.
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Invalid or expired token
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string

    # ── Public ────────────────────────────────────────────────────────────────
...
    RegisterRequest:
      type: object
      required:
        - full_name
        - phone
      properties:
        full_name:
          type: string
          example: John Doe
        phone:
          type: string
          example: "+12125551234"
        email:
          type: string
          format: email
          example: john@example.com
        hotel_id:
          type: integer
          example: 1
          description: ID from GET /hotels
        shuttle_opt_in:
          type: boolean
          default: true
        arrival_airline:
          type: string
          example: AA
        arrival_flight_number:
          type: string
          example: AA123
        arrival_datetime:
          type: string
          format: date-time
          example: "2026-06-11T14:30:00-05:00"
        departure_airline:
          type: string
          example: DL
        departure_flight_number:
          type: string
          example: DL456
        departure_datetime:
          type: string
          format: date-time
          example: "2026-06-14T18:00:00-05:00"
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        btl_code:
          type: string
          example: BTL-001
        message:
          type: string
    UpdateFlightRequest:
      type: object
      required:
        - btl_code
        - direction
      properties:
        btl_code:
          type: string
          example: BTL-001
        direction:
          type: string
          enum: [arrival, departure]
        airline:
          type: string
          example: AA
        flight_number:
          type: string
          example: AA123
        submitted_datetime:
          type: string
          format: date-time
          example: "2026-06-11T14:30:00-05:00"
    FlightInfo:
      type: object
      nullable: true
      properties:
        airline:
          type: string
        flight_number:
          type: string
        submitted_datetime:
          type: string
          format: date-time
        live_eta:
          type: string
          format: date-time
          nullable: true
        flight_status:
          type: string
          enum: [on_time, delayed, landed, cancelled, unknown]
        delay_mins:
          type: integer
        polling_active:
          type: boolean
        leg4_pickup_from:
          type: string
          enum: [church, hotel]
          nullable: true
    RunSummary:
      type: object
      properties:
        run_id:
          type: string
        run_type:
          type: string
        direction:
          type: string
          enum: [to_church, to_hotel, to_airport]
        depart_time:
          type: string
          example: "08:00"
        status:
          type: string
        vehicle_label:
          type: string
          nullable: true
        driver_name:
          type: string
          nullable: true
    ParticipantStatus:
      type: object
      properties:
        participant:
          type: object
          properties:
            btl_code:
              type: string
            full_name:
              type: string
            phone:
              type: string
            email:
              type: string
            status:
              type: string
              enum: [registered, confirmed, arrived, cancelled]
            needs_attention:
              type: boolean
            shuttle_opt_in:
              type: boolean
            hotel:
              type: object
              nullable: true
              properties:
                hotel_name:
                  type: string
                pickup_address:
                  type: string
        arrival:
          $ref: '#/components/schemas/FlightInfo'
        departure:
          $ref: '#/components/schemas/FlightInfo'
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RunSummary'
        generated_at:
          type: string
          format: date-time
...
    ParticipantSummary:
      type: object
      properties:
        id:
          type: integer
        btl_code:
          type: string
          example: BTL-001
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [registered, confirmed, arrived, cancelled]
        needs_attention:
          type: boolean
        attention_reason:
          type: string
          nullable: true
        shuttle_opt_in:
          type: boolean
        hotel:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            hotel_name:
              type: string
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    ParticipantListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantSummary'
        total_elements:
          type: integer
        total_pages:
          type: integer
        page:
          type: integer
        size:
          type: integer
...
paths:
...
  /register:
    post:
      tags: [Public]
      summary: Register a conference participant
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              full_name: John Doe
              phone: "+12125551234"
              email: john@example.com
              hotel_id: 1
              shuttle_opt_in: true
              arrival_airline: AA
              arrival_flight_number: AA123
              arrival_datetime: "2026-06-11T14:30:00-05:00"
              departure_airline: DL
              departure_flight_number: DL456
              departure_datetime: "2026-06-14T18:00:00-05:00"
      responses:
        "200":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                success: true
                btl_code: BTL-001
                message: Registration successful
  /update-flight:
    post:
      tags: [Public]
      summary: Submit or update a participant's flight details
      operationId: updateFlight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFlightRequest'
            example:
              btl_code: BTL-001
              direction: arrival
              airline: AA
              flight_number: AA123
              submitted_datetime: "2026-06-11T14:30:00-05:00"
      responses:
        "200":
          description: Flight updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /participant-status:
    get:
      tags: [Public]
      summary: Look up participant status by BTL code
      operationId: participantStatus
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            example: BTL-001
      responses:
        "200":
          description: Participant status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantStatus'
        "400":
          description: Participant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /twilio-webhook:
    post:
      tags: [Public]
      summary: Inbound SMS webhook (called by Twilio)
      operationId: twilioWebhook
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                From:
                  type: string
                Body:
                  type: string
      responses:
        "200":
          description: TwiML acknowledgement
          content:
            text/xml:
              schema:
                type: string

  # ── Admin — Auth ─────────────────────────────────────────────────────────────
...
  /admin/participants:
    get:
      tags: [Admin]
      summary: List participants (paginated, filterable)
      operationId: listParticipants
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [registered, confirmed, arrived, cancelled]
        - name: hotel_id
          in: query
          schema:
            type: integer
        - name: needs_attention
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          description: Full-text search across name, email, phone
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated participant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantListResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /admin/participants/{btlCode}:
    get:
      tags: [Admin]
      summary: Get a single participant by BTL code
      operationId: getParticipant
      security:
        - bearerAuth: []
      parameters:
        - name: btlCode
          in: path
          required: true
          schema:
            type: string
            example: BTL-001
      responses:
        "200":
          description: Participant detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipantSummary'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /admin/participants/{id}/hotel:
    patch:
      tags: [Admin]
      summary: Update participant hotel assignment
      operationId: updateParticipantHotel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hotel_id:
                  type: integer
            example:
              hotel_id: 2
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
...